Add the meat of a YAML-based budgetary system
The function "load_budget" provided in budget.py will load the budget
from the given directory.

The budget consists of a tree of YAML files.  Each YAML file
represents an item in the budget, and contains the cost, as well as a
short and long description of the item.

Costs are simple expressions, and sympy is used to evaluate these.
The variables specified in config.yaml are used when these expressions
are evaluated.  Intepretation of variable values is left until budget
items are parsed, which might allow variables to contain expressions
referencing other variables :)

Proper schema checking of the YAML files is not yet supported --
mainly because there isn't yet a schema checker library that is
generally in distro repositories.