Alter the budget API to be tree based
After looking at the budget, it seems that being able to
programmatically interact with it as a tree is useful.  This patch
modifies load_budget() to return a BudgetTree object.  BudgetTree
objects have children that are either other BudgetTrees or
BudgetItems.  Thus a tree is formed.

The old list-like behaviour can be achieved with a simple
modification.  The walk() member function of each BudgetTree object
generates all the BudgetItems that exist in that entire subtree.

So any code that was:

l = budget.load_budget("./")
for i in l:
    ...

Can be replaced with:

l = budget.load_budget("./"):
for i in l.walk():
    ...