Added a case for the beagleboard XMs
The XMs have no mounting holes, and therefore would be difficult to
attach to the scaffold, the case makes this connection easier and also
protects the XM, especially from shorts through the scaffold.

The cost estimate is based on it being bigger than the motor board
case, however no more than the power board case which is much bigger
and more complex.