Add a 'consumable' property to each budget item
It's incredibly useful to be able to categorise the budget items by
whether they represent a cost that, once spent, will only benefit SR
once or multiple times.

For example, cable ties are a consumable because they will be used
once and then we have to buy more.  Whereas a power board is not a
consumable because we continue to use it for years.