Add a tool for verifying that the budget is below the limit
As changes are made we need to ensure that the budget does not exceed
the amount that we have applied for (and will hopefully have...).
This tool checks that the total is below the max (which is the current
total), and outputs some useful messages, as well as returning useful
error codes.