Increase the emergency shipping fund
According to trac tickets[1] we spent £540.64 shipping kit to teams at
the start of the year. We also now need to spend £38.20 to ship a
replacement power board to EMM.

This patch covers the initial shipping costs, the cost of EMM's
replacement power board and gives us £71 for future shipping expenses.

[1] #929, #932, #933, #934

Modification accepted in this thread:
http://thread.gmane.org/gmane.science.robotics.srobo.general/1037