Update the cost of the arena carpet tiles.
The price previously quoted involved an offer that is no longer
available, and was also not for enough tiles. While we like to think
that the arena is 8*8, it's usually just larger than this. Allowing
an 8.5m square should cover this and give us enough spare should
some get damaged corners.