Add a budget item for a PPL licence to play music
Playing copyrighted works in public spaces (i.e. non-domestic) in the
UK requires a PPL licence.  I phoned PPL earlier to find out what we
needed to do.  They told me how much it'll cost.  This patch adds a
budget line that deals with this cost.

I checked with them whether there would be some pre-existing licence
that SUSU or the University has that we could use.  Apparently this is
not the case, as the SUSU has a different type of licence which is
in some way limited to specific events and/or tunes.

http://thread.gmane.org/gmane.science.robotics.srobo.general/1407