Reallocate hardboard budget to pirate ships
We have several times more hardboard than we actually need for the
competition, as we are no longer using it for the floor of the arena.

There's no budget line for the pirate ships currently.  This patch
reallocates the funds that we had for the hardboard to the parts for
the pirate ships.

[budget-mod] thread:
http://thread.gmane.org/gmane.science.robotics.srobo.general/1442
(Note it was approved during the time the SC vote count was reduced to 3).