Reassign the generator budget to cabling
We are now reasonably sure we don't need a generator for Garden
Court.  This patch reallocates the funds from the generator to the
hire of both audio and power cables, and the purchase of some RCDs for
use with those cables.

I don't know what we would have done if we had hired a generator... I
mean there'd be no cabling to connect to it with :-O

Approved in:
http://thread.gmane.org/gmane.science.robotics.srobo.general/1734