Assign funds to the parts for the bucket barrier
Rather obviously, we need money to create the bucket barrier.  We
determined that we didn't need to buy any additional cable ties (other
than the ones I've included in the arena fixings budget entry), so
this patch removes that, and assigns some of the result to the bucket
barrier.

Accepted in this thread:
http://article.gmane.org/gmane.science.robotics.srobo.general/1739