Migrate to using the budget library from tools.git
Since we're now interested in doing things like diffing between
versions of the budget, the format needs to stay relatively constant
throughout.  So it no longer makes sense to keep the budget API within
budget.git.  The API has been migrated to tools.git.

This patch keeps __init__.py so that things that use budget.git as a
submodule will continue to work for the time being.