Move info about incoming funds to funds-in.yaml
The `check` utility is shortly to move into tools.git.  This patch
separates the checking logic from the data required to do so.  Now,
money that has come in is stored in funds-in.yaml.

Note that we don't need to track incoming funds as hard as we do
outgoing funds -- because we generally always like incoming funds!