Split and reserve some sr2013 motor board funds
As now detailed in sr2014/kits/reserved-motors.yaml, we're not
building the "new" USB motor board in sr2012, but instead hope to
build it in sr2013.  This patch reduces the money allocated to the
batch that'll be manufactured in sr2012, and reserves the remainder
for later merging with the latest funding instalment.