Allocate more funds to purchasing kit boxes
After the last few days of vault rearrangement, it has become clear
that the only way we are going to be able to manage the increasing
quantities of stuff in our inventory is by having more boxes to store
it in.  At this point in time, we have *just* enough boxes to store
everything we have.  Come kickstart, we will need to ship more than
the number of boxes we currently have out to schools.  We need to buy
more boxes than just the number of new teams that we are getting
because we will still need to store stuff after kickstart.

This patch:
 - Corrects the kit box line to be about 18 litre boxes, as this is
   the size of the kit box that we use.
 - Increases the number of boxes we will purchase from 14 to 40.  This
   will allow us to continue to store things in an orderly fashion.