Free up definitely unused team budgets from SR2012
There were only 20 teams in SR2012, but there are still team budgets
allocated for 24.  This patch frees the budget currently allocated to
those 4 non-existent teams.

At the point of writing, £528.28 of this £1800 line has been spent.
At some point in the future, more of this budget line will be able to
be freed, but only after some kind of deadline for claims has been
served to teams.