Increase the SR2013 motor budget
We've had a quote back from Garner Osborne for the assembly of the
motor boards.  It's larger than previously budgeted.

The reason that it's larger is that the previous estimate was for 40
motor boards.  However, at the time of sending off for the quote, we
needed to create 48 motor boards to bring the number of motor boards
in known/certain locations in the inventory up to our SR2013 target.

So this patch increases the budget allocated to SR2013 motor boards.