Free up budget from various places
This frees up a large chunk of money to allow the mech kit budget lines
to be cherry picked from the moto-app-2012 branch.

The food, prizes and writebands lines have been reduced to their actual
spend. The team-budget line has been reduced to be closer to the actual
spend, with some capacity in case another team claims (although this is
now very unlikely). The clothing budget line has been reduced based on
an estimate I made for getting 30 t-shirts (at £7.50 a piece) and 10
polo shirts (at £10 a piece).