Copy sr2013/competition from moto-app-2012 branch
The sponsorship installment from the Motorola Foundation for SR2013-14
was smaller than anticipated.  This means that the branch that
contains the changes that describe our application (moto-app-2012) is
currently too large, and so cannot be merged right now.

At the time of writing this, it is 35 days until the SR2013
competition.  Money needs to be spent now on this event, and
significantly changing spending plans on this now would not be
terrifically wise.  Furthermore, the budget required for that
event is about 10% of the total budget, so changing its allocation
isn't going to result in savings that are going to bridge the gap in
funds.

Therefore, it makes sense to just copy sr2013/competition across from
the moto-app-2012 into master, and increment funds-in.yaml by the
total value of just those lines.  This is exactly what this patch does.

The sr2013/competition files that this patch injects are from
bb265885f7e4c824f3e6d0c08f806c8a66b5fb48 (which is the current HEAD of
the moto-app-2012 branch).

 What will happen when the moto-app-2012 branch is sorted?
===========================================================

When the moto-app-2012 branch has been reduced so that it fits the
available money, it will need to be merged into the master branch.
This is a simple case of:

% git checkout moto-app-2012
% git merge master
... adjust funds-in.yaml to the full amount and commit it ...
% git checkout master
% git merge moto-app-2012