Use ceiling() instead of ceil() in config.yaml
With sympy 0.7.2, the ceil() function no longer exists.  According to
some people I spoke to in the sympy IRC channel, the fact that ceil()
worked prior to 0.7.2 was actually an accident.  The correct function
to use is ceiling().

The error message that people were getting with 0.7.2 was highly
confusing due to another bug in sympy, which has been registered
upstream: http://code.google.com/p/sympy/issues/detail?id=3839

Change-Id: Ia548fda45f646a05fb0cf66cd79ee6c1be2ede6e