Revert "Consolidate the case costs into one budget line"
This reverts commit a93e5ff03e7aa6ac2bc510607d542168cc7cf779,
which was originally proposed in this list thread:
http://thread.gmane.org/gmane.science.robotics.srobo.general/2498
(it was pre-gerrit).

For an unwritten reason, this commit replaced three budget lines with
one [1].  There wasn't any need to merge them into one thing, as in
the purchasing ticket #1325 the cost that is associated with each of
those lines is perfectly clear.  (It *would* make sense if the company
we were purchasing things off didn't provide a breakdown of the cost
of these cases.  But they did, and the commit message should have said
such a thing anyway.)

*Before* this budget mod was generated, #1325 was registered against
the three budget lines that it removes.  There are currently three
entries in spending.git that reference those three budget lines.  The
easiest way out of this mess is to revert this budget mod that didn't
need to happen in the first place.

[1] sr2013/kits/{motor,power,servo}/cases with sr2013/kits/cases