Increase sr2012/competition/arena/barrier
At the time that this line was closed, an entry in spending.git
(2013/04/23/comp-barrier.yaml) that was supposed to reference it was
referencing an invalid budget line (it had been typed incorrectly).
This was fixed in spending.git 5be3b34b2d, and so this line needs to
have its value set to the correct value.  This patch does that.

A note for the SC:

  The latest version of budget-check now performs additional checks if
  you feed it the location of a spending.git repository using the -s
  switch -- e.g: `sr budget-check -s ~/mybees/spending`

  When provided with spending.git, as well as checking that the budget
  is below maximum, it will check:

   - All budget lines that spending.git references exist.  So if
     you're running budget-check to verify a proposed budget change,
     this will ensure that a budget line that's already had spends
     made against it isn't being deleted.

   - Closed lines are at exactly the correct value for the amount
     that's been spent against them.

   - All other lines have had less than 1.1 times their allocated cost
     spent against them.