Increase SR2014 battery budget
We're going to replace all the batteries we ship at teams with ones
from Hobby King.  The ones we had last year were of quite low quality,
and ~30% of them have already failed.  It seems likely that many more
of our batteries will fail during this year.  Furthermore, we're
changing connectors.

This doesn't mean that we're binning all of our old ones.  We can keep
hold of them and use the ones that survive at the competition (after
modding them to have new connectors).

Change-Id: I208d9626db62feae38d7043eb2a060ef99c739fe