Reduce budget allocated to SR2014 servo boards
At the time of our last funding application, we hatched a plan to ship
a USB servo board in SR2014.  However, we are somewhat behind on this
plan (an SR USB servo board does not currently exist).  We're now
planning on shipping a USB servo board in SR2015, and shipping out
SRIC servo boards in SR2014.

This means that we only need to manufacture 23 servo boards for
SR2014, rather than 81.  This frees a significant amount of budget.

Somewhat obviously, this means that we will need to find funds in the
near future to manufacture the USB servo boards for SR2015.  However,
freeing up this budget will allow us to fix a bunch of issues that we
have with the SR2014 kit budget that we have right now (#1723, #1725,
and #1726 to name a few).  Also note that our budget does not
currently extend to SR2015, so this is a reasonably valid action.

Change-Id: I53840c9929faa1e79f44c45ec3726e4d00d2d330