Augment SR2014 usb hub budget to get more ports
The SR2014 kit will require 6 USB ports (see #1724 for more details).
Our kit currently has a 4-port hub in it, which obviously does not
provide enough ports.

At the moment, budget is only allocated for purchasing one 4-port
hubs for each of the new kits for SR2014.

Exactly what we're going to do about this is undecided at the moment.
However, what's clear is that this budget line needs to be
significantly larger.

So this patch increases the line to be for a 7-port USB hub from a
reputable manufacturer (D-Link).  Hopefully we'll be able to reduce
this line over the next month or so as we research into USB hubs
more.

Change-Id: If62aa533f0bc31d91a4275877b37e0bcd4846787