Add config variables for SR2015
These assume that we are continuing our current trajectory of
multiplying the number of teams we have by 1.5.  81 teams in SR2015
:-O

Additionally, for SR2015, the magic factor determining how much spare
kit we have is reduced to 1.125 times from 1.25 times.  This seems
reasonable and fits with the amount of breakage that we have
experienced throughout the last year.

Change-Id: I099368022c4d03ed279b55cf285421b9f63d2c99