Add entries for arena display hire and beaglebones
With a growing number of arenas, we need more displays.

For SR2012, we purchased a set of 6 displays and BeagleBoard XMs to
suit.  This totalled £1840, or £306 per display.  By hiring the
displays, and using a BeagleBone Black, this cost is reduced to about
£100 per display.

By hiring, we avoid having to store so much stuff, and can have larger
displays too.

Change-Id: I09bec0ca46352d2751f772c35064ecd03a9032a7