Reallocate SUSU venu hire.
SUSU have elected to charge £850 for the venue hire: this is based on
4 days charged at £212.50 per day, which is the rate they're charging
the University itself to run the graduation ceremony in the same space.

Apparently the previous cost estimate was based on the belief that
we were lacking in sponsors, or were otherwise not capable of paying the
(somewhat subsidised) commercial rate.

Change-Id: Iea04e1cd1032a815b458bf6432183aad0434cf1c