Reallocate power/servo board costs based on quotes
I've got a quote for the PCB+assembly of both the power and servo
boards. I've also got a quote for all of the components for the power
and servo boards. As we're purchasing the components ourselves I've
split the servo board back out into separate budget lines.

This also takes into account that we're not manufacturing any more dev
power and servo boards due to this being the last year in which they
will be used.

This frees up around £790.

Change-Id: I98519a3925072dab4d9d84eb85f62b30af9d6000