Retroactively increase sr2013 prizes budget.
This patch stems from ticket #1613, where the purchase of the
certificates for SR2013 occurred. Due to some accidental extra VAT, the
cost of the certificates rose more than expected, and over the budgeted
amount. At the time, I took the 'too late; ship it; wrangle later'
approach, and the purchase went forward (and they were awesome
certificates).

However, the "wrangle it later" didn't occur, mostly because the amount
recorded in spending.git didn't change after the details in the ticket
changed (my fault), and wasn't noticed in the post-competition budget
cleanup. This, then, is the much belated wrangling, which allocates
about £50 from the existing free funds to repay the rest of the
certificates purchase.

Keeping trac and spending.git in sync is an unsolved problem, that
probably will be fine with sufficient diligence (and lack of stress).

The spending.git fixing change is g:761.

Change-Id: Iccbb1e1566568448b009a1e59a6303bdca675ed6