Increase sr2014 battery power cable budget
We only have 40 battery cables in our inventory at the moment, which
is significantly fewer than the 58 we thought we had at the time of
generating this budget line.  This patch bumps up the budget for
buying new cables to accommodate for this annoyance.  It also updates
the camcon price (gone up by 4p a connector)

Change-Id: Ia7f78978030885e8f68221ea2f5ee391d397a75f