Increase USB hub budget to cover SR2015 too
By buying 200 USB hubs now, we can cover what we need for SR2014 as
well as SR2015.  Buying them in this quantity will reduce the amount
of paperwork and hassle that we have to endure in purchasing them from
EntaTech in the future.

This patch increases the budget for USB hubs from £1628.19 to
£2140.80.

Change-Id: Ie1367a45c52bc2771a40920b186e1ba74cf1ec31