Add budget for SR2014 bubble wrap
We need to package up kits, and this patch adds a line for the bubble
wrap for this.

This uses the same estimation technique as last year, which turned out
to be reasonably accurate.  It assumes we need to fill 60% of the
volume of the kit boxes that we ship out with bubble wrap.

The small bubble wrap is for wrapping around each board.

Change-Id: I33f7d20090a71f1bf45d1c9cbcc01104eec5d119