Increase battery and charger budget for shipping
Our initial estimate of price didn't include the shipping from
Hobbyking.  This is because we were originally hoping to be able to
buy them from their consumer site etc.

In the purchasing ticket, the current spend is £1571.38.  This is
expected to increase to £1924.28 once VAT and duty have been paid.
This is 1.286 times more than we originally anticipated.

This patch increases the relevant parts of the battery and charger
budgets by this factor to spread the shipping out between them.

Change-Id: Idf607d7d53fd115eda12ac443db0d8b9ab9324ca