Reallocate some money for a screwdriver in mech kits
The screws in the screw sheilds are quite small and, based upon feedback
in previous years, people seem to struggle finding small screw drivers.

These are also useful for use at Kickstart, as teams will be using the
kit on the day and will need to be able to attach things to the screw
sheilds.

Change-Id: I1eacadb87770928e4638927a72556db1383af05c