Bump lawyer fee.
The invoice for this came in, and it's more than originally estimated. I
suspect that their previous assertion that they would charge 'up to
$blah pounds' only applied to the initial estimate of £600, and the
subsequent disclaimer's £300 estimate was unbounded.

A minute-by-minute breakdown of what it cost is also available, I'll
attach it to the next spending ticket against this line.

Either way, it's another £211.20 on top of what we've already paid them,
and we're not going to win an argument with lawyers about whether
they've done what we asked them to.

Change-Id: I3ccbb7df30d8e2d2106dbc813a9448249928c79a