Spread battery budget cost between budget lines
This redistributes the cost of ticket 1880 (where ~all the battery
equipment was bought) between the relevant budget lines. As a result,
£26.56 is freed up from the budget.

It's almost impossible to distribute these costs or break them down in
a sane manner (see: the ticket), or at least it'd require large
quantities of time. Therefore I've settled on just having the total
amount spent in the budget with no detailed breakdown; the historical
reader can read #1880 if they want to know what truly happened.

(This resolves over-budget distribution of spending in spending.git).

Change-Id: I7fa833e3ce0540dfe104f810da2a38aeff119016