Remove obsolete __init__.py and example.py
The budget is really only a place for data, not code.

The __init__.py is a hangover from a long while back when it also
contained the API for accessing it.  This is now in tools.git, and
things should be using that.  So if this breaks those things, then
they should be fixed to use tools.git's library.

Change-Id: Ie49afc5ad870fe830641534683164816b185923b