Add SR2014 competition-related items
This patch adds budget for the SR2014 competition.  It does this by
adding a loan that Jeremy and I are making to SR.

This is a temporary loan from Jeremy and I, which we expect to be paid
back promptly after our Motorola Foundation funding has reached our
bank account.  (Current indications from the SUSU are that this will
be about the 15th of April).

For reference, an easy way of viewing this (if you've cherry-picked
it) is: `sr budget-diff -t origin/master`