Adjust/increase production-related budget lines
Bennellick Engineering Ltd provided the power, lighting and cabling
for the event.  This patch increases the budget lines accordingly.

Yup, this is more expensive than in previous years, and it's more
expensive than the estimates that we had before.  However, as noted in
those budget lines, most of those estimates were stabs in the dark
(accurate estimates take several days to generate!).

Change-Id: I934bd462ac92a3e2efa430bdcf3c20f0b680244b