Correct 2014 competition sound to being consumable.
This avoids a bunch of potential issues in the future. Since this
file is likely to be used as a reference (even if not copy/pasted)
for 2015, it seems important that it be correct to avoid propogation
of the error. Additionally, when budgeting for sound next time around,
it's going to be rather annoying to look at this line and think
(incorrectly) that we can re-use the hire again, which is wrong.

Change-Id: I1ff605d3830450c3b8e62acb66bf1f7b5f1aafda