Increase SSL cert line to allow for wildcard cert
It is looking incredibly likely that we are soon going to need to serve
things from more than one server.  Our HTTPS cert currently only covers
www.studentrobotics.org, and in order to support multiple servers, it
needs to cover any subdomain of studentrobotics.org.  This patch
increases the budget for the certificate to allow for this.

Our current certificate expires on the 11th of August this year, and we
are planning on purchasing the new one on the 1st of August.

Change-Id: I5fc6a080ff00f1eb41386192dc2c5eefef33005c