Reallocate free tablet funds to wifi dongles
The wifi dongles that we had planned on getting are now not
available (no longer manufactured).  The replacement dongles that we can
get on short notice are £8.70 each + VAT, which is more than the £8 that
was budgeted for them.  This patch closes the tablets line (as we've
got the tablets now) and reallocates the freed budget to the wifi
dongles.

Change-Id: Id89da5610641e08e7c574caa73e53ebd8ec89a38